House UI and Hover Fully Working
Added improved assets for the UI, as well as some 3rd party button asssets. I also created scripts to manage the house UI and what is shown, i also created a hover script for the character windows on the house which is fully working for Mabel. For Rose i will require Rose's pixel asset and to update the text when needed. To make her window functional i will need to attach another script to to a trigger box to make it be seen. Worked on Scenes: Rose_Garden_Elizabeth_5th and Bedroom_Elizabeth_5th.